+++ src/components/navigation/side-nav/side-nav.scss
@@ -3,7 +3,6 @@ nav {
 }
 .MagentaA11y__side-nav-container {
   display: flex;
-  position: relative;
 
   .MagentaA11y__side-nav {
     --side-nav-background-color: var(--color-sidenav);
