+++ src/hooks/useKeyboardNavigation.ts
@@ -1,0 +1,22 @@
+import { useEffect, useState } from 'react';
+
+export const useKeyboardNavigation = () => {
+  const [isKeyboardNavigation, setIsKeyboardNavigation] = useState(false);
+
+  useEffect(() => {
+    const handleKeyDown = () => setIsKeyboardNavigation(true);
+    const handleMouseDown = () => setIsKeyboardNavigation(false);
+
+    window.addEventListener('keydown', handleKeyDown);
+    window.addEventListener('mousedown', handleMouseDown);
+    window.addEventListener('touchstart', handleMouseDown);
+
+    return () => {
+      window.removeEventListener('keydown', handleKeyDown);
+      window.removeEventListener('mousedown', handleMouseDown);
+      window.removeEventListener('touchstart', handleMouseDown);
+    };
+  }, []);
+
+  return isKeyboardNavigation;
+};
